Checkpoint: Removed Simple* implementations of the org.apache.nifi.registry.flow.* interfaces and changed the interfaces to classes. They are simply POJOs so no need for an Interface. Having the interface caused Jackson to not be able to properly marshal and unmarshal responses
