+++ nifi-api/src/main/java/org/apache/nifi/registry/flow/FlowRegistryBucket.java
@@ -16,15 +16,51 @@
  */
 package org.apache.nifi.registry.flow;
 
-public interface FlowRegistryBucket {
+public class FlowRegistryBucket {
+    private String identifier;
+    private String name;
+    private String description;
+    private long createdTimestamp;
+    private FlowRegistryPermissions permissions;
 
-    String getIdentifier();
+    public String getIdentifier() {
+        return identifier;
+    }
 
-    String getName();
+    public String getName() {
+        return name;
+    }
 
-    String getDescription();
+    public String getDescription() {
+        return description;
+    }
 
-    long getCreatedTimestamp();
+    public long getCreatedTimestamp() {
+        return createdTimestamp;
+    }
 
-    FlowRegistryPermissions getPermissions();
+    public FlowRegistryPermissions getPermissions() {
+        return permissions;
+    }
+
+    public void setIdentifier(final String identifier) {
+        this.identifier = identifier;
+    }
+
+    public void setName(final String name) {
+        this.name = name;
+    }
+
+    public void setDescription(final String description) {
+        this.description = description;
+    }
+
+    public void setCreatedTimestamp(final long createdTimestamp) {
+        this.createdTimestamp = createdTimestamp;
+    }
+
+    public void setPermissions(final FlowRegistryPermissions permissions) {
+        this.permissions = permissions;
+    }
+
 }
