--- nifi-commons/nifi-registry-client-utils/src/main/java/org/apache/nifi/registry/flow/SimpleRegisteredFlowSnapshotMetadata.java
@@ -1,80 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.nifi.registry.flow;
-
-public class SimpleRegisteredFlowSnapshotMetadata implements RegisteredFlowSnapshotMetadata {
-    private String bucketIdentifier;
-    private String flowIdentifier;
-    private int version;
-    private long timestamp;
-    private String author;
-    private String comments;
-
-    @Override
-    public String getBucketIdentifier() {
-        return bucketIdentifier;
-    }
-
-    @Override
-    public String getFlowIdentifier() {
-        return flowIdentifier;
-    }
-
-    @Override
-    public int getVersion() {
-        return version;
-    }
-
-    @Override
-    public long getTimestamp() {
-        return timestamp;
-    }
-
-    @Override
-    public String getAuthor() {
-        return author;
-    }
-
-    @Override
-    public String getComments() {
-        return comments;
-    }
-
-    public void setBucketIdentifier(String bucketIdentifier) {
-        this.bucketIdentifier = bucketIdentifier;
-    }
-
-    public void setFlowIdentifier(String flowIdentifier) {
-        this.flowIdentifier = flowIdentifier;
-    }
-
-    public void setVersion(int version) {
-        this.version = version;
-    }
-
-    public void setTimestamp(long timestamp) {
-        this.timestamp = timestamp;
-    }
-
-    public void setAuthor(String author) {
-        this.author = author;
-    }
-
-    public void setComments(String comments) {
-        this.comments = comments;
-    }
-}
