+++ nifi-nar-bundles/nifi-flow-registry-client-bundle/nifi-flow-registry-client-services/pom.xml
@@ -38,11 +38,6 @@
             <version>1.18.0-SNAPSHOT</version>
         </dependency>
         <dependency>
-            <groupId>org.apache.nifi</groupId>
-            <artifactId>nifi-registry-client-utils</artifactId>
-            <version>1.18.0-SNAPSHOT</version>
-        </dependency>
-        <dependency>
             <groupId>org.apache.nifi.registry</groupId>
             <artifactId>nifi-registry-client</artifactId>
             <version>1.18.0-SNAPSHOT</version>
